requirements.txt
# Core ML frameworks
torch>=2.0.0
transformers>=4.44.0
accelerate>=0.34.0

# Quantization libraries
autoawq>=0.2.0
auto-gptq>=0.7.0
bitsandbytes>=0.41.0
exllamav2>=0.0.11
hqq>=0.1.0

# Datasets and evaluation
datasets>=2.19.0
scipy>=1.10.0
scikit-learn>=1.3.0

# RAG components
sentence-transformers>=2.2.0
chromadb>=0.4.0
rank-bm25>=0.2.2

# Metrics
rouge-score>=0.1.2
nltk>=3.8

# Utilities
PyYAML>=6.0
tqdm>=4.65.0
pandas>=2.0.0
numpy>=1.24.0

# lm-eval-harness
lm-eval>=0.4.4

# Optional - for additional formats
PyPDF2>=3.0.0
openpyxl>=3.1.0  # For Excel export